Delayed transition support & implementation for the TimerInteraction for AkkaBaker (#1468)
Changed the TimerInteraction actor to be a generic DelayedTransitionActor.
The processInstance now knows TransitionDelayed & DelayedTransitionFired next to the regular TransitionFiredEvent. This allows for transition executions to be owned by other actors (in this case the DelayedTransitionActor).
Implemented persistence for the DelayedTransitionActor.

Ensured the SensoryEventResponseHandler is stopped after the first event if it is only interested in the Received functionality.